// src/types/GameUnits.d.ts
export type UnitName =
    "militia" |
    "sword" |
    "slinger" |
    "archer" |
    "hoplite" |
    "rider" |
    "chariot" |
    "catapult" |
    "minotaur" |
    "manticore" |
    "zyklop" |
    "harpy" |
    "medusa" |
    "centaur" |
    "pegasus" |
    "cerberus" |
    "fury" |
    "griffin" |
    "calydonian_boar" |
    "satyr" |
    "spartoi" |
    "ladon" |
    "godsent" |
    "big_transporter" |
    "bireme" |
    "attack_ship" |
    "demolition_ship" |
    "small_transporter" |
    "trireme" |
    "colonize_ship" |
    "sea_monster" |
    "siren"